Biome configuration redo Part 1
Resetting all Eras and Eternity to a base to work with. This stage is incomplete as all extra biomes are disabled in all Eras at this point.

Stronghold and village generation settings will be consistent across all Eras and Eternity, but of course not all the biomes will be present in each era. They also happen to be the defaults.

Prototypes for Tesseral and Wasteland terrain gen will be present as part of this pass.