Biome Pass Part 7: Information
A "half-step" in time compared to all the other Eras, this marks the "present day" and the turning point where history can wander one way or the other.

Flickers of both Tesseral and Wasteland gen are mixed in with a tweaked Mechanized worldgen.

Origin Valley biomes are unusually common in this period as attempts to avert environmental disaster are generating these biomes as experiments.