Build 086 Phase 1:
Updates all mods in change log EXCEPT Pam's Harvestcraft as i need to delete the original file and upload a new one in quick succession. For now I am just staging the files. This also updates the CoFH ore generation files to fix the internal change made to the configs and here the vanilla replacement switch is.